+++ sensors.go
@@ -8,7 +8,7 @@ import (
 )
 
 // Liest die CPU-Temperatur des Raspberry Pi aus und gibt diese als String zurück
-func cpuTemp(tempChan chan float64) {
+func cpuTemp(tempChan chan float32) {
 	// Kommando auf dem Raspberry Pi ausführen und 0 in den Channel schreiben, falls ein Fehler auftritt
 	cmd := exec.Command("vcgencmd", "measure_temp")
 	out, err := cmd.Output()
@@ -23,7 +23,7 @@ func cpuTemp(tempChan chan float64) {
 	tempStr = strings.TrimSpace(tempStr)
 	tempStr = strings.Trim(tempStr, "temp='C")
 	temp, _ := strconv.ParseFloat(tempStr, 32)
-	tempChan <- temp
+	tempChan <- float32(temp)
 }
 
 // Liest die CPU-Taktfrequenz des Raspberry Pi aus und gibt diese als String zurück
